Added boost
This one is a pain in the ass. They have their own builder thing that it builds itself and then you use it to build the others, except it's obtuse and even ChatGPT is confused on how to use it. And if you try and install to /usr/local it fails and doesn't say why. I wound up just having it build to a directory, complete with header files, and then just copy *that* to /usr/local.

It's *probably* easier if you're not trying to do a Universal 2 thing like I'm doing but thems the breaks.