Added openssl@3
Because (I believe) this library requires/insists on a platform triplet (i.e., "darwin64-arm64-cc", it has to be build separate, installed, and lipo'd later. It's possible I'm just not being imaginative enough and it could work with one of the existing make scripts but for now I'm giving it its own thing. I may factor this out later.